requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.5
django-cors-headers==3.7.0
django-currentuser==0.5.2
django-debug-toolbar==3.2
django-notifications==0.1.dev0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.1.0
pkg-resources==0.0.0
pycodestyle==2.6.0
PyJWT==2.0.1
pytz==2020.5
sqlparse==0.4.1
toml==0.10.2
